move top level rete modules into nodes
